agregado boton "volver" en otras paginas, eliminado el limpiar el formulario de creacion de vehiculo al haber un error, agregado un texto al crear caracteristicas "no hay caracteristicas creadas"

+++ src/Routes/Categories.jsx
@@ -4,15 +4,16 @@ import AdministracionPhoneError from '../Components/Phone Error/AdministracionPh
 import { Link } from 'react-router-dom'
 import { useContextGlobal } from "../Context/GlobalContext";
 import { routes } from "../utils/routes";
+import BackButton from '../Components/BackButton/BackButton';
 
 const CreateCategories = () => {
     const { state, getToken, dispatch } = useContextGlobal();
     const token = getToken();
     const [render, setRender] = useState(true);
-    const [ brandInputValue, setBrandInputValue ] = useState("");
-    const [ modelInputValue, setModelInputValue ] = useState("");
-    const [ typeInputValue, setTypeInputValue ] = useState("");
-    const [ featureInputValue, setFeatureInputValue ] = useState("");
+    const [brandInputValue, setBrandInputValue] = useState("");
+    const [modelInputValue, setModelInputValue] = useState("");
+    const [typeInputValue, setTypeInputValue] = useState("");
+    const [featureInputValue, setFeatureInputValue] = useState("");
 
     const formReference = useRef(null);
     const handleBrandInput = (e) => {
@@ -60,39 +61,40 @@ const CreateCategories = () => {
                 'Authorization': "Basic " + token,
             },
         }).then(res => {
-                alert("Categoría creada con éxito")
-                switch (category) {
-                    case 'brand':
-                        dispatch({ type: 'SET_LIST_BRAND', payload: [...state.brand, res.data] });
-                        break;
-                    case 'type':
-                        dispatch({ type: 'SET_LIST_TYPE', payload: [...state.type, res.data] });
-                        break;
-                    case 'model':
-                        dispatch({ type: 'SET_LIST_MODEL', payload: [...state.model, res.data] });
-                        break;
-                    default:
-                        break;                
-                }
-            }).catch(err => {
-                console.log(err)
-            })
+            alert("Categoría creada con éxito")
+            switch (category) {
+                case 'brand':
+                    dispatch({ type: 'SET_LIST_BRAND', payload: [...state.brand, res.data] });
+                    break;
+                case 'type':
+                    dispatch({ type: 'SET_LIST_TYPE', payload: [...state.type, res.data] });
+                    break;
+                case 'model':
+                    dispatch({ type: 'SET_LIST_MODEL', payload: [...state.model, res.data] });
+                    break;
+                default:
+                    break;
+            }
+        }).catch(err => {
+            console.log(err)
+        })
     }
 
     function deleteCategory(e, string) {
         e.preventDefault()
-        const categoryToDelete = state[`${string}`].find( 
+        if (!confirm("¿Estás seguro que deseas eliminar esta categoría?")) return;
+        const categoryToDelete = state[`${string}`].find(
             (value) => {
-                return ( e.target.elements[0].value === value.name );
+                return (e.target.elements[0].value === value.name);
             }
         );
 
-        let aux = Array.from( string );
-        let categoryIdField = ( aux.map( (letter, index) => { return ( index === 0 ) ? letter.toUpperCase() : letter } ) )
+        let aux = Array.from(string);
+        let categoryIdField = (aux.map((letter, index) => { return (index === 0) ? letter.toUpperCase() : letter }))
         categoryIdField = categoryIdField.join("");
-        categoryIdField = "id" + categoryIdField; 
+        categoryIdField = "id" + categoryIdField;
         axios.delete(
-            `${ routes.url_base }/${ string }/delete/${ categoryToDelete[`${categoryIdField}`] }`,
+            `${routes.url_base}/${string}/delete/${categoryToDelete[`${categoryIdField}`]}`,
             {
                 headers: {
                     'Content-Type': 'application/json',
@@ -100,30 +102,30 @@ const CreateCategories = () => {
                 },
             }
         ).then(res => {
-                alert("Categoria eliminada")
-                    dispatch({ 
-                    type: `SET_LIST_${string.toUpperCase()}`, 
-                    payload: ( state[`${string}`].filter( (element) => element[`${categoryIdField}`] !== categoryToDelete[`${categoryIdField}`]) ) 
-                });
-                console.log( state[`${string}`] );
-            }).catch(err => {
-                alert("Categoria en uso: No es posible, eliminar una categoría que esté siendo usada!");
-            })
+            alert("Categoria eliminada")
+            dispatch({
+                type: `SET_LIST_${string.toUpperCase()}`,
+                payload: (state[`${string}`].filter((element) => element[`${categoryIdField}`] !== categoryToDelete[`${categoryIdField}`]))
+            });
+            console.log(state[`${string}`]);
+        }).catch(err => {
+            alert("Categoria en uso: No es posible, eliminar una categoría que esté siendo usada!");
+        })
     }
 
     function createFeature(e) {
         e.preventDefault()
         if (featureInputValue === "") return alert("Campos vacios");
         if (featureInputValue[0] === " ") return alert("Campos vacios");
-        setFeatureInputValue(""); 
-        if ( (state.feature.find(feature => feature.name === featureInputValue))) {
-            setFeatureInputValue(""); 
-            return alert('Ya existe esa categoría');  
+        setFeatureInputValue("");
+        if ((state.feature.find(feature => feature.name === featureInputValue))) {
+            setFeatureInputValue("");
+            return alert('Ya existe esa categoría');
         }
-        axios.post(`${routes.url_base}/feature`, 
-            { 
-                name: featureInputValue 
-            }, 
+        axios.post(`${routes.url_base}/feature`,
+            {
+                name: featureInputValue
+            },
             {
                 headers: {
                     'Content-Type': 'application/json',
@@ -131,23 +133,23 @@ const CreateCategories = () => {
                 },
             }
         ).then(res => {
-                alert("Característica creada con éxito");
- 
-                setFeatureInputValue(""); 
-                dispatch({ type: 'SET_LIST_FEATURES', payload: [...state.feature, res.data] });
-            }).catch(err => {
+            alert("Característica creada con éxito");
+
+            setFeatureInputValue("");
+            dispatch({ type: 'SET_LIST_FEATURES', payload: [...state.feature, res.data] });
+        }).catch(err => {
         })
     }
 
     const deleteFeature = (e) => {
         e.preventDefault();
-        const featureToDelete = state.feature.find( 
+        if (!confirm("¿Estás seguro que deseas eliminar esta característica?")) return;
+        const featureToDelete = state.feature.find(
             (value) => {
-                return ( e.target.elements[0].value === value.name );
+                return (e.target.elements[0].value === value.name);
             }
         );
-        axios.delete(`${routes.url_base}/feature/delete/${featureToDelete.idFeature}`, 
-
+        axios.delete(`${routes.url_base}/feature/delete/${featureToDelete.idFeature}`,
             {
                 headers: {
                     'Content-Type': 'application/json',
@@ -155,23 +157,22 @@ const CreateCategories = () => {
                 },
             }
         ).then(res => {
-                alert("Característica eliminada")
-                dispatch({ type: 'SET_LIST_FEATURES', payload: ( state.feature.filter( (feature) => feature.idFeature !== featureToDelete.idFeature)  ) });
-            }).catch(err => {
-                alert("Característica en uso: No es posible, eliminar una característica que esté siendo usada!");
-            })
+            alert("Característica eliminada")
+            dispatch({ type: 'SET_LIST_FEATURES', payload: (state.feature.filter((feature) => feature.idFeature !== featureToDelete.idFeature)) });
+        }).catch(err => {
+            alert("Característica en uso: No es posible eliminar una característica que esté siendo usada!");
+        })
     }
 
     setTimeout(() => {
-        setRender(false)  
+        setRender(false)
     }, 180);
 
     return (
         <>
-            {render ?  <p className="loader">Loading....</p> :  
-
+            {render ? <p className="loader">Loading....</p> :
                 <div className='create__categories__container'>
-                    <Link to={`/administracion`}><h3>Volver</h3></Link>
+                    <BackButton />
                     <h2 className='title__admin'>Administración</h2>
                     <div className="administracion__funciones">
                         <div className="administracion__categories">
@@ -180,7 +181,7 @@ const CreateCategories = () => {
                                     <h3 id='create'>Crear categorías</h3>
                                     <div>
                                         <h4>Marca: </h4>
-                                        <input type="text" id='brand-input' value={brandInputValue} onInput={handleBrandInput}/>
+                                        <input type="text" id='brand-input' value={brandInputValue} onInput={handleBrandInput} />
                                         <button onClick={(e) => {
                                             const input = document.getElementById('brand-input');
                                             const brand = input.value;
@@ -190,8 +191,8 @@ const CreateCategories = () => {
                                     </div>
                                     <div>
                                         <h4>Modelo: </h4>
-                                        <input type="text" id='model-input' value={modelInputValue} onInput={handleModelInput}/>
-                                        <button onClick={ (e) => {
+                                        <input type="text" id='model-input' value={modelInputValue} onInput={handleModelInput} />
+                                        <button onClick={(e) => {
                                             const input = document.getElementById('model-input');
                                             const model = input.value;
                                             if (state.model.includes(model)) return alert("Ya existe esa categoría")
@@ -201,7 +202,7 @@ const CreateCategories = () => {
                                     </div>
                                     <div>
                                         <h4>Tipo: </h4>
-                                        <input type="text" id='type-input' value={typeInputValue} onInput={handleTypeInput}/>
+                                        <input type="text" id='type-input' value={typeInputValue} onInput={handleTypeInput} />
                                         <button onClick={(e) => {
                                             const input = document.getElementById('type-input');
                                             const type = input.value;
@@ -262,7 +263,7 @@ const CreateCategories = () => {
                                     <h3 id='create'>Crear característica</h3>
                                     <div>
                                         <h4>Característica: </h4>
-                                        <input type="text" id='feature-input' value={featureInputValue} onInput={handleFeatureInput}/>
+                                        <input type="text" id='feature-input' value={featureInputValue} onInput={handleFeatureInput} />
                                         <button onClick={createFeature}>Crear</button>
                                     </div>
